Add stats section + fix name problem
There was a problem where the riot LoL api deviated from the convention. Kai'Sa got transformed into KaiSa, following the convention, but to get specific Kai'Sa data, the name needed suddenly to be transformed to Kaisa. The same problem occured with Nunu & Willump, Nunu was needed. Now it is fixed